requirements.txt
beautifulsoup4==4.13.4
python-dotenv==1.0.1
requests==2.32.3
